Adição, Listagem e remoção de produtos

+++ src/main/resources/db/migration/V001__create-table-product.sql
@@ -2,7 +2,7 @@ create table product(
 	id bigint not null auto_increment,
 	name varchar(60) not null,
 	price decimal(5,2) not null,
-	score tinyint not null,
+	score int not null,
 	image varchar(60),
 
 	primary key(id)
